+++ src/main/kotlin/net/casual/championships/CasualChampionships.kt
@@ -21,12 +21,13 @@ import net.casual.championships.config.CasualConfig
 import net.casual.championships.lobby.minigame.modules.LobbyData
 import net.casual.championships.lobby.minigame.modules.LobbyParkourData
 import net.casual.championships.minigame.CasualMinigameManager
-import net.casual.championships.minigame.duel.CasualDuelArenas
+import net.casual.championships.minigame.duel.DuelArenas
 import net.casual.championships.resources.CasualResourcePackHost
 import net.casual.championships.sync.CasualDatabaseSyncService
 import net.casual.championships.sync.CasualNoopSyncService
 import net.casual.championships.sync.CasualSyncService
 import net.casual.championships.uhc.minigame.UHCMinigameFactory
+import net.casual.championships.util.CasualComponentUtils
 import net.casual.database.CasualDatabase
 import net.fabricmc.api.DedicatedServerModInitializer
 import net.fabricmc.loader.api.FabricLoader
@@ -51,12 +52,7 @@ object CasualChampionships: DedicatedServerModInitializer {
         CasualUtils.logger.info("Starting CasualChampionships... Version: ${container.metadata.version}")
 
         CasualResourcePackHost.registerEvents()
-        CasualDuelArenas.registerEvents()
-
-        // TODO: We should register these somewhere else...
-        UHCMinigameFactory.register(MinigameRegistries.MINIGAME_FACTORY)
-        LobbyData.register(MinigameRegistries.MINIGAME_DATA_MODULE_PROVIDER)
-        LobbyParkourData.register(MinigameRegistries.MINIGAME_DATA_MODULE_PROVIDER)
+        DuelArenas.registerEvents()
 
         this.minigames.registerEvents(GlobalEventHandler.Server)
         GlobalEventHandler.Server.register<ServerStartEvent>(priority = 10_000, listener = ::onServerStart)
@@ -66,28 +62,18 @@ object CasualChampionships: DedicatedServerModInitializer {
     fun reload(server: MinecraftServer) {
         this.config = CasualConfig.read()
 
-        CasualDuelArenas.reload(server)
+        DuelArenas.reload(server)
         this.minigames.reload(server)
 
         this.reloadSyncService()
     }
 
-    fun getMessageOfTheDay(): Component {
-        return Component {
-            val cc = literal("Casual Championships".convertCasing(TitleCase, SmallCapsTitleCase)).bold().color(0xFFAC1C)
-            val title = literal("\uD83D\uDDE1").yellow() + " " + cc + " " + literal("\uD83C\uDFF9").yellow()
-            val subtitle = literal("   be prepared".toSmallCaps()).lime() + wrap() +
-                literal(" ◆ ").white() + literal("let the chaos ensue    ".toSmallCaps()).lime()
-            empty() + literal("╔═════", 0x009BFF, 0x80CDFF) + title + literal("═════╗", 0x80CDFF, 0x009BFF) + nl +
-                literal("╚══", 0x009BFF, 0x33AFFF) + subtitle + literal("══╝", 0x33AFFF, 0x009BFF)
-        }
-    }
-
     private fun onServerStart(event: ServerStartEvent) {
         val (server) = event
-        server.setMessageOfTheDay(this.getMessageOfTheDay())
         this.minigames.load(server)
         this.reloadSyncService()
+
+        server.setMessageOfTheDay(CasualComponentUtils.getMessageOfTheDay())
     }
 
     private fun onServerRegisterCommand(event: ServerRegisterCommandEvent) {
