+++ src/main/kotlin/net/casual/championships/minigame/CasualMinigameManager.kt
@@ -43,6 +43,7 @@ import net.casual.championships.duel.minigame.DuelMinigame
 import net.casual.championships.lobby.minigame.LobbyMinigame
 import net.casual.championships.minigame.event.EventConfiguration
 import net.casual.championships.minigame.event.EventState
+import net.casual.championships.minigame.lobby.LobbyMinigames
 import net.casual.championships.minigame.lobby.LobbySidebar
 import net.casual.championships.resources.CasualResourcePackHost
 import net.casual.championships.sync.data.SyncableParticipants
@@ -226,7 +227,7 @@ class CasualMinigameManager(
     }
 
     private fun createLobby(server: MinecraftServer): LobbyMinigame {
-        val lobby = LobbyMinigame.create(this.config.lobby, this::minigame, MinigameCreationContext(server))
+        val lobby = LobbyMinigames.create(this.config.lobby, this::minigame, MinigameCreationContext(server))
         lobby.resources.add(CasualResourcePackHost.createResourcesFromPacks { lobby.getAdditionalPacks() })
         this.modifyLobbyMinigame(lobby)
         return lobby
